feat(auth): merge login and register forms with previous versions
