+++ app/(auth)/login/page.tsx
@@ -1,11 +1,9 @@
-import { LoginForm as AuthLoginForm } from "@/components/auth/login-form";
-import { LoginForm } from "@/components/login-form";
+import { LoginForm } from "@/components/auth/login-form";
 
 export default function Page() {
 	return (
 		<div className="flex min-h-svh flex-col items-center justify-center gap-6 bg-background p-6 md:p-10">
 			<div className="w-full max-w-sm">
-				<AuthLoginForm />
 				<LoginForm />
 			</div>
 		</div>
