+++ app/(auth)/register/page.tsx
@@ -1,12 +1,10 @@
 import { RegisterForm } from "@/components/auth/register-form";
-import { SignupForm } from "@/components/signup-form";
 
 export default function Page() {
 	return (
 		<div className="flex min-h-svh flex-col items-center justify-center gap-6 bg-background p-6 md:p-10">
 			<div className="w-full max-w-sm">
 				<RegisterForm />
-				<SignupForm />
 			</div>
 		</div>
 	);
