+++ components/auth/register-form.tsx
@@ -1,20 +1,35 @@
 "use client";
 
-import { zodResolver } from "@hookform/resolvers/zod";
+import { standardSchemaResolver } from "@hookform/resolvers/standard-schema";
+import Link from "next/link";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
+import { OAuthButton } from "@/components/auth/oauth-button";
+import { AppIcon } from "@/components/icons/app-icon";
+import { Button } from "@/components/ui/button";
+import {
+	Field,
+	FieldDescription,
+	FieldGroup,
+	FieldLabel,
+	FieldSeparator,
+} from "@/components/ui/field";
+import { Input } from "@/components/ui/input";
 import { authClient } from "@/lib/auth-client";
-import { signInWith42 } from "./oauth42";
+import { cn } from "@/lib/utils";
 import { type RegisterFormInput, registerSchema } from "./schemas";
 
-export function RegisterForm() {
+export function RegisterForm({
+	className,
+	...props
+}: React.ComponentProps<"div">) {
 	const [apiError, setApiError] = useState<Record<string, string>>({});
 	const {
 		register,
 		handleSubmit,
 		formState: { errors, isSubmitting },
 	} = useForm<RegisterFormInput>({
-		resolver: zodResolver(registerSchema),
+		resolver: standardSchemaResolver(registerSchema),
 	});
 
 	const onSubmit = async (data: RegisterFormInput) => {
@@ -40,91 +55,93 @@ export function RegisterForm() {
 		}
 	};
 
-	const onOAuth42Register = async () => {
-		await signInWith42(setApiError);
-	};
-
 	return (
-		<form onSubmit={handleSubmit(onSubmit)} className="grid gap-4">
-			{apiError.general && (
-				<div className="rounded-md bg-red-50 p-3 text-sm text-red-800">
-					{apiError.general}
-				</div>
-			)}
+		<div className={cn("flex flex-col gap-6", className)} {...props}>
+			<form onSubmit={handleSubmit(onSubmit)}>
+				<FieldGroup>
+					<div className="flex flex-col items-center gap-2 text-center">
+						<Link
+							href="/"
+							className="flex flex-col items-center gap-2 font-medium"
+						>
+							<div className="flex size-8 items-center justify-center rounded-md">
+								<AppIcon className="size-6" />
+							</div>
+							<span className="sr-only">ChatSimulator</span>
+						</Link>
+						<h1 className="text-xl font-bold">
+							Create your ChatSimulator account
+						</h1>
+						<FieldDescription>
+							Already have an account? <Link href="/login">Sign in</Link>
+						</FieldDescription>
+					</div>
 
-			<div className="grid gap-1">
-				<label htmlFor="email" className="text-sm font-medium">
-					Email
-				</label>
-				<input
-					{...register("email")}
-					type="email"
-					id="email"
-					placeholder="Enter your email"
-					className="rounded-md border border-gray-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none"
-					disabled={isSubmitting}
-				/>
-				{(errors.email || apiError.email) && (
-					<p className="text-xs text-red-600">
-						{errors.email?.message || apiError.email}
-					</p>
-				)}
-			</div>
+					{apiError.general && (
+						<div className="rounded-md bg-red-50 p-3 text-sm text-red-800">
+							{apiError.general}
+						</div>
+					)}
 
-			<div className="grid gap-1">
-				<label htmlFor="password" className="text-sm font-medium">
-					Password
-				</label>
-				<input
-					{...register("password")}
-					type="password"
-					id="password"
-					placeholder="Enter your password"
-					className="rounded-md border border-gray-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none"
-					disabled={isSubmitting}
-				/>
-				{(errors.password || apiError.password) && (
-					<p className="text-xs text-red-600">
-						{errors.password?.message || apiError.password}
-					</p>
-				)}
-			</div>
+					<Field>
+						<FieldLabel htmlFor="email">Email</FieldLabel>
+						<Input
+							{...register("email")}
+							id="email"
+							type="email"
+							placeholder="m@example.com"
+							disabled={isSubmitting}
+						/>
+						{(errors.email || apiError.email) && (
+							<p className="text-xs text-red-600">
+								{errors.email?.message || apiError.email}
+							</p>
+						)}
+					</Field>
 
-			<div className="grid gap-1">
-				<label htmlFor="confirmPassword" className="text-sm font-medium">
-					Confirm Password
-				</label>
-				<input
-					{...register("confirmPassword")}
-					type="password"
-					id="confirmPassword"
-					placeholder="Confirm your password"
-					className="rounded-md border border-gray-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none"
-					disabled={isSubmitting}
-				/>
-				{errors.confirmPassword && (
-					<p className="text-xs text-red-600">
-						{errors.confirmPassword.message}
-					</p>
-				)}
-			</div>
+					<Field>
+						<FieldLabel htmlFor="password">Password</FieldLabel>
+						<Input
+							{...register("password")}
+							id="password"
+							type="password"
+							placeholder="Enter your password"
+							disabled={isSubmitting}
+						/>
+						{errors.password && (
+							<p className="text-xs text-red-600">{errors.password.message}</p>
+						)}
+					</Field>
 
-			<button
-				type="submit"
-				disabled={isSubmitting}
-				className="mt-2 rounded-md bg-blue-600 px-4 py-2 text-sm font-medium text-white hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed"
-			>
-				{isSubmitting ? "Creating account..." : "Register"}
-			</button>
+					<Field>
+						<FieldLabel htmlFor="confirmPassword">Confirm Password</FieldLabel>
+						<Input
+							{...register("confirmPassword")}
+							id="confirmPassword"
+							type="password"
+							placeholder="Confirm your password"
+							disabled={isSubmitting}
+						/>
+						{errors.confirmPassword && (
+							<p className="text-xs text-red-600">
+								{errors.confirmPassword.message}
+							</p>
+						)}
+					</Field>
 
-			<button
-				type="button"
-				onClick={onOAuth42Register}
-				disabled={isSubmitting}
-				className="rounded-md border border-gray-300 px-4 py-2 text-sm font-medium text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-			>
-				Continue with 42
-			</button>
-		</form>
+					<Button type="submit" disabled={isSubmitting} className="w-full">
+						{isSubmitting ? "Creating account..." : "Create Account"}
+					</Button>
+
+					<FieldSeparator>Or</FieldSeparator>
+					<OAuthButton setApiError={setApiError} disabled={isSubmitting} />
+				</FieldGroup>
+			</form>
+			<FieldDescription className="px-6 text-center">
+				By clicking continue, you agree to our{" "}
+				<Link href="/terms">Terms of Service</Link> and{" "}
+				<Link href="/privacy">Privacy Policy</Link>.
+			</FieldDescription>
+		</div>
 	);
 }
